roman.py tests/test_roman.py README.md: Test for LXXXIX. Fixed bugs.
LXXXIX has four Xs and is a legitimate roman numeral.

Bugs in the rules in README.md about how many times a letter
may appear in a roman numeral were fixed to limit how many
time a letter may appear _consecutively_.

roman.py was changed to limit consecutive letters,
not total number of a letter in a roman numeral.