roman.py: Refactored: Works, but much more complicated, so do not use.
I am leaving this commit here as a monument to a failed experiment.

I thought the way that I did things before was too complicated,
that it would be easier if I did the conversion backwards,
so I tried it.

It works, and the code is smaller,
but it is so complicated, it makes my head hurt.
I used to be proud of small tight tricky code like this, but not anymore.

Also, the way I got to it was not the incremental TDD way.
The code of this commit sucks, but I still do not like
the previous working code, so I am going to start over
with another technique that is hopefully simpler.