roman.py: Roman.__str__: Start with most significant digit.
The old code started with the least significant digit.
The nice thing about the old way was that it only converted
the digits that needed to be converted.
The drawback to the old way were the two reversed() operations,
which were slightly tricky.

The new way is more straightFORWARD, starting with the most significant digit.
This makes it easier to read.
The drawback of the new way is that it always converts all possible digits,
even when the number has fewer significant digits.

I chose the new way for its readability.