src/roman.c: get_value_of_roman_numeral(): Check validity with regular expression.
This is the first step of using a regular expression to
check Roman numerals for validity instead of a bunch of code.
The regex was generated by the Python version,
then I had to add \\ in front of all '(', '|', ')', and '?'
characters.

I have not removed the old validity checking code yet,
so the regex stuff is redundant for now.

Removing the old validity checks will be the next steps.