src/roman.c: get_value_of_roman_numeral(): Polish.
Using a regular expression to figure out most of the validity of
a Roman numeral worked out great. The code is so much simpler now.

I would not use the regex technique for tiny little microcontrollers,
like an 8051 based microcontroller, because the regex library would
probably take too much code space, but for something running on
Ubuntu, there is plenty of RAM.

For a tiny microcontroller, I would probably use the techique of
tricky logic that uses little RAM,
such as in commit addf67b 2017-11-13 23:10:03 -0500.

There is an interesting edge case of an empty string.
An empty string matches the regex, although it is not a Roman numeral.
The loop gracefully does nothing, so sum stays at zero.
That is returned as the valid for a bad roman numeral.
It is one of those situations where bad input (emtpy string)
returns the correct value even though that case is not checked for.